+++ src/routes/FilterCreateRoute.js
@@ -1,8 +1,8 @@
 import { omit } from 'lodash';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
 
 import {
+  NoPermissionMessage,
   useOkapiKyMutation,
   useOkapiKyQuery,
 } from '@folio/stripes-leipzig-components';
@@ -52,7 +52,7 @@ const FilterCreateRoute = ({ history, location }) => {
   };
 
   if (!hasPerms) {
-    return <div><FormattedMessage id="ui-finc-select.noPermission" /></div>;
+    return <NoPermissionMessage />;
   }
 
   return (